https://harsh-simon-says.netlify.app/
ClicK here To play : https://harsh-simon-says.netlify.app/

Simon Says is a children's game for three or more players. One player takes the role of "Simon" and issues instructions (usually physical actions such as "jump in the air" or "stick out your tongue") to the other players, which should be followed only when prefaced with the phrase "Simon says". Players are eliminated from the game by either following instructions that are not immediately preceded by the phrase, or by failing to follow an instruction which does include the phrase "Simon says". It is the ability to distinguish between genuine and fake commands, rather than physical ability, that usually matters in the game; in most cases, the action just needs to be attempted.